+++ R/EWmethod.R
@@ -110,6 +110,13 @@ EWMethod <- function(initialSample, b = n, increases = FALSE)
     stop("Parameter b should be integer value and > 0")
   }
 
+  # checking the validity of increases
+
+  if(!is.logical(increases))
+  {
+    stop("Parameter increases should have logical value")
+  }
+
 
   # check form of the initial sample