+++ R/GeneratorNExpUU.R
@@ -93,6 +93,13 @@ GeneratorNExpUU <- function(n, mu, sigma, lambda, b, c, increases = FALSE, ...)
     stop("Parameter n should be integer value and > 1")
   }
 
+  # checking mu parameter
+
+  if(!is.double(mu))
+  {
+    stop("Parameter mu should be double value")
+  }
+
   # checking sigma parameter
 
   if(!is.double(sigma) | sigma < 0)
@@ -119,6 +126,13 @@ GeneratorNExpUU <- function(n, mu, sigma, lambda, b, c, increases = FALSE, ...)
   if(!is.double(c) | c < 0)
   {
     stop("Parameter c should be double value and > 0")
+  }
+
+  # checking the validity of increases
+
+  if(!is.logical(increases))
+  {
+    stop("Parameter increases should have logical value")
   }