+++ R/SEResamplingMean.R
@@ -139,6 +139,13 @@ SEResamplingMean <- function(initialSample, resamplingMethod=ClassicalBootstrap,
     stop("Parameter resamplingMethod should be a proper name of the resampling method")
   }
 
+  # checking the validity of increases
+
+  if(!is.logical(increases))
+  {
+    stop("Parameter increases should have logical value")
+  }
+
 
   # matrix for means