+++ R/VAmethod.R
@@ -42,7 +42,7 @@
 #'
 #' @family resampling functions
 #'
-#' @seealso @seealso \code{\link{ClassicalBootstrap}},
+#' @seealso \code{\link{ClassicalBootstrap}},
 #' \code{\link{EWMethod}} for the EW method, \code{\link{VAFMethod}} for the VAF method,
 #' \code{\link{VAAMethod}} for the VAA method, \code{\link{DMethod}} for the d method, \code{\link{WMethod}} for the w method
 #'
@@ -111,12 +111,20 @@ VAMethod <- function(initialSample, b = n, increases = FALSE)
     stop("Parameter b should be integer value and > 0")
   }
 
+  # checking the validity of increases
+
+  if(!is.logical(increases))
+  {
+    stop("Parameter increases should have logical value")
+  }
+
   # check form of the initial sample
 
   if(increases)
   {
     initialSample <- TransformFromIncreases(initialSample)
   }
+
 
   # checking consistency of fuzzy numbers